% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_dataset.R
\name{Dataset}
\alias{Dataset}
\title{R6 Class representing a person.}
\description{
This is a new description line
}
\details{
A person has a name and a hair color.

This is a third line.
}
\section{Usage}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{d = Dataset$new(studies="GSE34126")
d$email_address = 'your_email@ex.com'
d$save()$start_processing()$wait_until_ready()$download()
}\if{html}{\out{</div>}}
}

\examples{

## ------------------------------------------------
## Method `Dataset$set_id`
## ------------------------------------------------

P <- Person("Ann", "black")
P$hair
P$set_hair("red")
P$hair
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Dataset-new}{\code{Dataset$new()}}
\item \href{#method-Dataset-run_pipeline}{\code{Dataset$run_pipeline()}}
\item \href{#method-Dataset-save}{\code{Dataset$save()}}
\item \href{#method-Dataset-start_processing}{\code{Dataset$start_processing()}}
\item \href{#method-Dataset-set_id}{\code{Dataset$set_id()}}
\item \href{#method-Dataset-update}{\code{Dataset$update()}}
\item \href{#method-Dataset-wait_until_ready}{\code{Dataset$wait_until_ready()}}
\item \href{#method-Dataset-download}{\code{Dataset$download()}}
\item \href{#method-Dataset-extract}{\code{Dataset$extract()}}
\item \href{#method-Dataset-create_summarized_experiment_from_extract}{\code{Dataset$create_summarized_experiment_from_extract()}}
\item \href{#method-Dataset-clone}{\code{Dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-new}{}}}
\subsection{Method \code{new()}}{
Create a new \link{Dataset}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(
  studies,
  email_address,
  quantile_normalize = NULL,
  quant_sf_only = NULL,
  svd_algorithm = NULL,
  scale_by = NULL,
  aggregate_by = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studies}}{a character() vector
of experiments or studies to include in the dataset (eg., GSE346126)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{\link[=Dataset$new]{Dataset$new()}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-run_pipeline"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-run_pipeline}{}}}
\subsection{Method \code{run_pipeline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$run_pipeline()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-save"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-start_processing"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-start_processing}{}}}
\subsection{Method \code{start_processing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$start_processing()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-set_id}{}}}
\subsection{Method \code{set_id()}}{
Change hair color.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$set_id(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New hair color.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{P <- Person("Ann", "black")
P$hair
P$set_hair("red")
P$hair
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-update"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-wait_until_ready"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-wait_until_ready}{}}}
\subsection{Method \code{wait_until_ready()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$wait_until_ready(seconds = 10)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-download"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-download}{}}}
\subsection{Method \code{download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$download(base_path = ".")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-extract"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$extract(base_path = ".")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-create_summarized_experiment_from_extract"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-create_summarized_experiment_from_extract}{}}}
\subsection{Method \code{create_summarized_experiment_from_extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$create_summarized_experiment_from_extract()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Dataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
