% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_page_browser.R
\name{dataset_page_browser}
\alias{dataset_page_browser}
\title{Open a webpage for the given dataset}
\usage{
dataset_page_browser(dataset)
}
\arguments{
\item{dataset}{The dataset ID or \code{rb_dataset} object}
}
\value{
the dataset page URL (invisibly)
}
\description{
This function will open a browser window to the RefineBio dataset page
for the given dataset. This is a convenience function for users who
want to quickly view the dataset page for a given dataset.
}
\examples{
\dontrun{
dataset_page_browser(dataset)
}

}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
