% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_page_browser.R,
%   R/dataset_sharing_link.R
\name{dataset_sharing_link}
\alias{dataset_sharing_link}
\title{Get a sharing link for the given dataset}
\usage{
dataset_sharing_link(dataset)

dataset_sharing_link(dataset)
}
\arguments{
\item{dataset}{The dataset ID or \code{rb_dataset} object}
}
\value{
A sharing link for the given dataset

A sharing link for the given dataset
}
\description{
This function will return a sharing link for the given dataset.
The link can be used to share the dataset with others.

This function will return a sharing link for the given dataset.
The link can be used to share the dataset with others.
}
\examples{
\dontrun{
dataset_sharing_link(dataset)
}

\dontrun{
dataset_sharing_link(dataset)
}

}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
