% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_datasets.R
\name{datastore_datasets}
\alias{datastore_datasets}
\title{List cached datasets'}
\usage{
datastore_datasets(base_path = datastore_get_path())
}
\arguments{
\item{base_path}{The path to the local datastore.}
}
\value{
A character vector of cached dataset IDs.
}
\description{
When you download a dataset from refine.bio, it is cached locally.
This function will list all of the cached datasets.
}
\examples{

test_datatore_path <- system.file(
  "testdata/datastore",
  package = "RefineBio"
)
list.files(test_datatore_path, recursive = TRUE)
datastore_datasets(test_datatore_path)

}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
