% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_path.R
\name{datastore_example_path}
\alias{datastore_example_path}
\title{An example datastore}
\usage{
datastore_example_path()
}
\description{
This is a small datastore that is used for testing.
}
\seealso{
Other Local datastore operations: 
\code{\link{datastore_get_path}()},
\code{\link{datastore_set_path}()}
}
\concept{Local datastore operations}
