% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_example_path.R,
%   R/datastore_get_path.R
\name{datastore_get_path}
\alias{datastore_get_path}
\title{Local datastore path settings}
\usage{
datastore_get_path()

datastore_get_path()
}
\value{
The path to the local datastore.

The path to the local datastore.
}
\description{
The refine.bio R package caches downloaded datasets locally.

The refine.bio R package caches downloaded datasets locally.
}
\details{
This function returns the path to the local datastore.

This function returns the path to the local datastore.
}
\seealso{
Other Local datastore operations: 
\code{\link{datastore_example_path}()},
\code{\link{datastore_set_path}()}

Other Local datastore operations: 
\code{\link{datastore_example_path}()},
\code{\link{datastore_set_path}()}
}
\concept{Local datastore operations}
