% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_path.R
\name{datastore_set_path}
\alias{datastore_set_path}
\title{Set the path to the refine.bio datastore}
\usage{
datastore_set_path(path)
}
\arguments{
\item{path}{\code{character(1)}
The path to the local datastore.}
}
\value{
\code{character(1)}
The path to the local datastore.
}
\description{
The refine.bio R package caches downloaded datasets locally.
}
\details{
This function sets the path to the local datastore.

Order of precedence:
\enumerate{
\item \code{datastore_path} argument
\item \code{refinebioc.datastore} option
\item \code{REFINEBIOC_DATASTORE} environment variable
}
}
\seealso{
Other Local datastore operations: 
\code{\link{datastore_example_path}()},
\code{\link{datastore_get_path}()}
}
\concept{Local datastore operations}
