% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_processing.R
\name{.process_json_result}
\alias{.process_json_result}
\title{process the httr json response}
\usage{
.process_json_result(result, encoding = "UTF-8", ...)
}
\arguments{
\item{result}{\link[httr:response]{httr::response} object with body assumed to be json}

\item{encoding}{character(1) the encoding passed to \code{\link[httr:content]{httr::content()}}}
}
\description{
process the httr json response
}
\keyword{internal}
