% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_state.R
\name{.rb_get_email_address}
\alias{.rb_get_email_address}
\title{internal function to get the email address for the current session}
\usage{
.rb_get_email_address()
}
\description{
internal function to get the email address for the current session
}
\keyword{internal}
