% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.refinebio_base}
\alias{.refinebio_base}
\title{refinebio url}
\format{
An object of class \code{character} of length 1.
}
\usage{
.refinebio_base
}
\description{
refinebio url
}
\keyword{datasets}
