% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_local_dataset.R
\name{extract_local_dataset}
\alias{extract_local_dataset}
\title{load a list of \code{SummarizedExperiment}s from an extracted RefineBio download}
\usage{
extract_local_dataset(basedir)
}
\arguments{
\item{basedir}{Path to the directory of an extracted RefineBio download.}
}
\value{
A list of \code{SummarizedExperiment}s representing the
experiments in the RefineBio downloaded dataset.
}
\description{
The download directory from RefineBio is a zip file containing a folder of
experiments. This function will load the experiments from the extracted
zip file and return a list of \code{SummarizedExperiment}s. In the case of a
single experiment, the list will contain a single element.
}
\details{
The returned list will be named based on the experiment ids in the download.
}
\examples{
# load the SummarizedExperiment package
# We'll need it later.
suppressPackageStartupMessages(library(SummarizedExperiment))

# use an example datastore
datastore_path <- datastore_example_path()

# Get the first dataset from the datastore
example_dataset <- datastore_datasets(datastore_path)[1]

# Get the path to the dataset
example_dataset_path <- file.path(datastore_path, example_dataset)

# Load the dataset
example_se_list <- extract_local_dataset(example_dataset_path)

# Get the first experiment
example_se <- example_se_list[[1]]

example_se

# Get the first experiment's metadata
str(metadata(example_se))

}
\author{
Alexander Ho and Sean Davis
}
