% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten a list}
\usage{
flatten_list(x)
}
\description{
Convert a list with multiple elements
into a flattened list keeping the names
of the original list
}
\examples{
a <- list(b = 1:5, d = LETTERS[1:3])

}
