% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiment_catalog.R
\name{get_experiment_catalog}
\alias{get_experiment_catalog}
\title{Complete list of all experiments in refine.bio}
\usage{
get_experiment_catalog(.pages = 1e+05)
}
\arguments{
\item{.pages}{\code{integer(1)} The number of pages
to retrieve (with default=\code{Inf}).
Generally, you'll want to ignore this;
it's only useful for testing.}
}
\value{
\code{tibble} A tibble of all experiments in refine.bio.

The tibble contains the following columns:
\itemize{
\item id: The ID of the experiment
\item title: The title of the experiment
\item publication_title: The title of the
publication associated with the experiment
\item description: A description of the experiment
\item accession_code: The accession code of the experiment
\item alternate_accession_code: An alternate accession code for the experiment
\item technology: The technology used to generate the data in the experiment
\item submitter_institution: The institution that submitted the experiment
\item has_publication: Whether the experiment
has a publication associated with it
\item publication_doi: The DOI of the publication associated with the experiment
\item publication_authors: The authors of
the publication associated with the experiment
\item pubmed_id: The PubMed ID of the publication associated with the experiment
\item source_first_published: The date the experiment was first published
\item num_total_samples: The total number of samples in the experiment
\item num_processed_samples: The number of samples
in the experiment that have been processed
\item num_downloadable_samples: The number of samples
in the experiment that are downloadable
\item platform_names: The names of the platforms used in the experiment
\item organism_names: The names of the organisms in the experiment
\item sample_metadata_fields: The types of metadata
available for the samples in the experiment
\item samples: a data.frame of sample metadata
}
}
\description{
Refine.bio experiments are collections
of samples that were processed together.
In practices, refine.bio experiments
are usually a GEO series or an SRA study.
This function returns a list of all experiments in refine.bio as a tibble.
}
\examples{
# this takes about 1-2 minutes to complete
# for the entire experiment listing
# depending on your internet connection
# You only need to do this once (or when you want
#   to update the results)

# get the first 2 pages of experiments for
# testing and demonstration purposes
experiments <- experiment_listing(.pages = 2)

# get all experiments
# experiments <- get_experiment_catalog()

head(experiments)
dim(experiments)
colnames(experiments)

}
\author{
\itemize{
\item Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
\item Alex Ho \href{mailto:alexander.2.ho@cuanschutz.edu}{alexander.2.ho@cuanschutz.edu}
}
}
\concept{experiments}
