% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_state.R
\name{get_rb_email_address}
\alias{get_rb_email_address}
\title{get the email address for the current session}
\usage{
get_rb_email_address(required = TRUE)
}
\arguments{
\item{required}{logical, whether to throw an
error if the email address is not set}
}
\description{
get the email address for the current session
}
\examples{
\dontrun{
get_rb_email_address()
}

}
