% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refinebio.R
\name{get_refinebio}
\alias{get_refinebio}
\title{Get data from RefineBio}
\usage{
get_refinebio(
  studies,
  email_address,
  quantile_normalize = NULL,
  quant_sf_only = NULL,
  svd_algorithm = NULL,
  scale_by = NULL,
  aggregate_by = NULL
)
}
\arguments{
\item{studies}{A list of study ids to download.}

\item{email_address}{The email address to send the download link to and
required to download the data.}

\item{quantile_normalize}{Whether to quantile normalize the data.}

\item{quant_sf_only}{Whether to only quantile normalize the signal fraction.}

\item{svd_algorithm}{The algorithm to use for SVD.}

\item{scale_by}{The scale factor to use for scaling the data.}

\item{aggregate_by}{The aggregation method to use for the data.}
}
\value{
A list of \link[SummaryExperiment:SummarizedExperiment]{SummaryExperiment::SummarizedExperiment} objects.
}
\description{
Get data from RefineBio
}
\examples{
\dontrun{
selist <- get_refinebio(c("GSE34126", "SRP113332"), email_address = "email@myemail.com")
}

}
