% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refinebio.R
\name{process_and_load_dataset}
\alias{process_and_load_dataset}
\title{Process and load a dataset}
\usage{
process_and_load_dataset(ds, base_path = datastore_get_path())
}
\arguments{
\item{ds}{The dataset ID or \code{rb_dataset} object}

\item{base_path}{The path to the directory where the dataset will be
downloaded and extracted.}
}
\value{
A list of \code{SummarizedExperiment}s representing the
experiments in the RefineBio downloaded dataset.
}
\description{
This function will process and load a dataset from RefineBio. It is
intended for use with the \code{rb_get_experiment} function. However,
it can be used with any dataset that has been registered by RefineBio.
This function will download, extract, and load the data from RefineBio
for the given dataset.
}
