% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_dataset.R
\name{rb_browse_dataset}
\alias{rb_browse_dataset}
\title{Open a webpage for the given dataset}
\usage{
rb_browse_dataset(dataset)
}
\arguments{
\item{dataset}{The dataset ID or \code{rb_dataset} object}
}
\description{
This function will open a browser window to the RefineBio dataset page
for the given dataset. This is a convenience function for users who
want to quickly view the dataset page for a given dataset.
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
