% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_datasets.R
\name{rb_cached_datasets}
\alias{rb_cached_datasets}
\title{List cached datasets'}
\usage{
rb_cached_datasets(base_path = rb_data_path())
}
\arguments{
\item{base_path}{The path to the directory where the datasets are cached.}
}
\value{
A character vector of cached dataset IDs.
}
\description{
When you download a dataset from refine.bio, it is cached locally.
This function will list all of the cached datasets.
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
