% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_compendia_list.R
\name{rb_compendia_list}
\alias{rb_compendia_list}
\title{List all CompendiaResults with filtering.}
\usage{
rb_compendia_list(
  primary_organism__name = NULL,
  compendium_version = NULL,
  quant_sf_only = NULL,
  result__id = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL,
  latest_version = NULL
)
}
\arguments{
\item{primary_organism__name}{string}

\item{compendium_version}{number}

\item{quant_sf_only}{boolean \code{True} for RNA-seq Sample Compendium results or \code{False} for quantile normalized.}

\item{result__id}{number}

\item{ordering}{string Which field to use when ordering the results.}

\item{limit}{integer Number of results to return per page.}

\item{offset}{integer The initial index from which to return the results.}

\item{latest_version}{boolean \code{True} will only return the highest \code{compendium_version} for each primary_organism.}
}
\description{
List all CompendiaResults with filtering.
}
\seealso{
Other compendia: 
\code{\link{rb_compendia_read}()}
}
\concept{compendia}
