% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_compendia_read.R
\name{rb_compendia_read}
\alias{rb_compendia_read}
\title{Get a specific Compendium Result}
\usage{
rb_compendia_read(id)
}
\arguments{
\item{id}{integer A unique integer value identifying this compendium result. (required)}
}
\description{
Get a specific Compendium Result
}
\seealso{
Other compendia: 
\code{\link{rb_compendia_list}()}
}
\concept{compendia}
