% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_computational_results_list.R
\name{rb_computational_results_list}
\alias{rb_computational_results_list}
\title{This lists all \code{ComputationalResult}. Each one contains meta-information about the output of a computer process. (Ex Salmon).}
\usage{
rb_computational_results_list(
  processor__id = NULL,
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{processor__id}{number}

\item{limit}{integer Number of results to return per page.}

\item{offset}{integer The initial index from which to return the results.}
}
\description{
This can return valid S3 urls if a valid \href{#tag/token}{token} is sent in the header \code{HTTP_API_KEY}.
}
\seealso{
Other computational_results: 
\code{\link{rb_computational_results_read}()}
}
\concept{computational_results}
