% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_computational_results_read.R
\name{rb_computational_results_read}
\alias{rb_computational_results_read}
\title{Retrieves a computational result by its ID}
\usage{
rb_computational_results_read(id)
}
\arguments{
\item{id}{integer A unique integer value identifying this computational result. (required)}
}
\description{
Retrieves a computational result by its ID
}
\seealso{
Other computational_results: 
\code{\link{rb_computational_results_list}()}
}
\concept{computational_results}
