% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_computed_files_list.R
\name{rb_computed_files_list}
\alias{rb_computed_files_list}
\title{ComputedFiles are representation of files created by refinebio processes.}
\usage{
rb_computed_files_list(
  id = NULL,
  samples = NULL,
  is_qn_target = NULL,
  is_smashable = NULL,
  is_qc = NULL,
  is_compendia = NULL,
  quant_sf_only = NULL,
  svd_algorithm = NULL,
  compendium_version = NULL,
  created_at = NULL,
  last_modified = NULL,
  result__id = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{id}{number}

\item{samples}{string}

\item{is_qn_target}{string}

\item{is_smashable}{string}

\item{is_qc}{string}

\item{is_compendia}{string}

\item{quant_sf_only}{string}

\item{svd_algorithm}{string}

\item{compendium_version}{number}

\item{created_at}{string}

\item{last_modified}{string}

\item{result__id}{number}

\item{ordering}{string Which field to use when ordering the results.}

\item{limit}{integer Number of results to return per page.}

\item{offset}{integer The initial index from which to return the results.}
}
\description{
It's possible to download each one of these files by providing a valid token. To
acquire and activate an API key see the documentation for the \href{#tag/token}{/token} endpoint.
When a valid token is provided the url will be sent back in the field \code{download_url}. Example:\if{html}{\out{<div class="sourceCode py">}}\preformatted{import requests
import json
headers = \{
    'Content-Type': 'application/json',
    'API-KEY': token_id # requested from /token
\}
requests.get('https://api.refine.bio/v1/computed_files/?id=5796866', \{\}, headers=headers)
}\if{html}{\out{</div>}}

This endpoint can also be used to fetch all the compendia files we have generated with:\preformatted{GET /computed_files?is_compendia=True&is_public=True
}
}
\seealso{
Other computed_files: 
\code{\link{rb_computed_files_read}()}
}
\concept{computed_files}
