% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_computed_files_read.R
\name{rb_computed_files_read}
\alias{rb_computed_files_read}
\title{Retrieves a computed file by its ID}
\usage{
rb_computed_files_read(id)
}
\arguments{
\item{id}{integer A unique integer value identifying this computed file. (required)}
}
\description{
Retrieves a computed file by its ID
}
\seealso{
Other computed_files: 
\code{\link{rb_computed_files_list}()}
}
\concept{computed_files}
