% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_dataset_create.R
\name{rb_dataset_create}
\alias{rb_dataset_create}
\title{View and modify a single Dataset.}
\usage{
rb_dataset_create(
  id = NULL,
  data = NULL,
  aggregate_by = NULL,
  scale_by = NULL,
  is_processing = NULL,
  is_processed = NULL,
  is_available = NULL,
  has_email = NULL,
  email_address = NULL,
  email_ccdl_ok = NULL,
  notify_me = NULL,
  expires_on = NULL,
  s3_bucket = NULL,
  s3_key = NULL,
  success = NULL,
  failure_reason = NULL,
  created_at = NULL,
  last_modified = NULL,
  start = NULL,
  size_in_bytes = NULL,
  sha1 = NULL,
  quantile_normalize = NULL,
  quant_sf_only = NULL,
  svd_algorithm = NULL
)
}
\arguments{
\item{id}{string}

\item{data}{object This is a dictionary where the keys are experiment accession codes and the values are lists with sample accession codes. Eg: \verb{\{'E-ABC-1': ['SAMP1', 'SAMP2']\}}}

\item{aggregate_by}{string Specifies how samples are \href{http://docs.refine.bio/en/latest/main_text.html#aggregations}{aggregated}.}

\item{scale_by}{string Specifies options for \href{http://docs.refine.bio/en/latest/main_text.html#transformations}{transformations}.}

\item{is_processing}{boolean}

\item{is_processed}{boolean}

\item{is_available}{boolean}

\item{has_email}{string}

\item{email_address}{string}

\item{email_ccdl_ok}{boolean}

\item{notify_me}{boolean}

\item{expires_on}{string}

\item{s3_bucket}{string}

\item{s3_key}{string}

\item{success}{boolean}

\item{failure_reason}{string}

\item{created_at}{string}

\item{last_modified}{string}

\item{start}{boolean}

\item{size_in_bytes}{integer Contains the size in bytes of the processed dataset.}

\item{sha1}{string}

\item{quantile_normalize}{boolean Part of the advanced options. Allows \href{http://docs.refine.bio/en/latest/faq.html#what-does-it-mean-to-skip-quantile-normalization-for-rna-seq-samples}{skipping quantile normalization} for RNA-Seq samples.}

\item{quant_sf_only}{boolean Include only quant.sf files in the generated dataset.}

\item{svd_algorithm}{string Specifies choice of SVD algorithm}
}
\description{
View and modify a single Dataset.
}
\seealso{
Other dataset: 
\code{\link{rb_dataset_read}()},
\code{\link{rb_dataset_update}()}
}
\concept{dataset}
