% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rb_dataset_download}
\alias{rb_dataset_download}
\title{Download an available dataset from refine.bio}
\usage{
rb_dataset_download(
  dataset,
  base_path = datastore_get_path(),
  overwrite = FALSE
)
}
\arguments{
\item{base_path}{The path to the directory where the dataset will be
downloaded and extracted.}

\item{overwrite}{If TRUE, overwrite the cached result.}

\item{ID}{The dataset ID}
}
\value{
The path to the downloaded zipfile.
}
\description{
Note that the dataset must be available before it can be downloaded.
}
