% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rb_dataset_extract}
\alias{rb_dataset_extract}
\title{Extract a downloaded dataset zipfile}
\usage{
rb_dataset_extract(
  dataset,
  base_path = datastore_get_path(),
  overwrite = FALSE
)
}
\arguments{
\item{base_path}{Path to the directory where the zipfile will be extracted.}

\item{overwrite}{If TRUE, overwrite the existing directory.}

\item{ID}{The dataset ID}
}
\value{
The path to the extracted directory.
}
\description{
Extract a downloaded dataset zipfile
}
