% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rb_dataset_load}
\alias{rb_dataset_load}
\title{Load a dataset from a refinebio download}
\usage{
rb_dataset_load(dataset, base_path = rb_data_path(), use_caching = TRUE)
}
\arguments{
\item{base_path}{The path to the directory where the dataset will be
downloaded and extracted.}

\item{ID}{The dataset ID}

\item{cache_result}{If TRUE, cache the result in the base_path.}

\item{overwrite}{If TRUE, overwrite the cached result.}
}
\value{
A list of \code{SummarizedExperiment}s representing the
experiments in the RefineBio downloaded dataset.
}
\description{
Load a dataset from a refinebio download
}
