% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_builder.R
\name{rb_dataset_loader}
\alias{rb_dataset_loader}
\title{load a list of \code{SummarizedExperiment}s from an extracted RefineBio download}
\usage{
rb_dataset_loader(basedir)
}
\arguments{
\item{basedir}{Path to the directory of an extracted RefineBio download.}
}
\value{
A list of \code{SummarizedExperiment}s representing the
experiments in the RefineBio downloaded dataset.
}
\description{
The download directory from RefineBio is a zip file containing a folder of
experiments. This function will load the experiments from the extracted
zip file and return a list of \code{SummarizedExperiment}s. In the case of a
single experiment, the list will contain a single element.
}
\details{
The returned list will be named based on the experiment ids in the download.
}
\author{
Alexander Ho and Sean Davis
}
\keyword{internal}
