% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_dataset_read.R
\name{rb_dataset_read}
\alias{rb_dataset_read}
\title{View a single Dataset.}
\usage{
rb_dataset_read(details = NULL)
}
\arguments{
\item{details}{boolean When set to \code{True}, additional fields will be included in the response with details about the experiments in the dataset. This is used mostly on the dataset page in www.refine.bio}
}
\description{
View a single Dataset.
}
\seealso{
Other dataset: 
\code{\link{rb_dataset_create}()},
\code{\link{rb_dataset_update}()}
}
\concept{dataset}
