% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rb_dataset_update_state}
\alias{rb_dataset_update_state}
\title{Get the state of a dataset}
\usage{
rb_dataset_update_state(dataset)
}
\arguments{
\item{dataset}{The dataset ID or \code{rb_dataset} object}
}
\value{
A \code{rb_dataset} object, now with updated state.
}
\description{
This function simply calls the API to get the state of a dataset.
}
