% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_state.R
\name{rb_email_address}
\alias{rb_email_address}
\title{get the email address for the current session}
\usage{
rb_email_address()
}
\description{
get the email address for the current session
}
\examples{
\dontrun{
get_rb_email_address()
}

}
