% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{rb_experiment_details}
\alias{rb_experiment_details}
\title{Get experiment data for a single experiment}
\usage{
rb_experiment_details(accession_code)
}
\arguments{
\item{accession_code}{The accession code of the experiment, e.g. 'SRP150473'}
}
\value{
A list with the following elements:
\itemize{
\item id: The ID of the experiment
\item title: The title of the experiment
\item description: A description of the experiment
\item annotations: A list of annotations for the experiment
\item samples: a data.frame of sample metadata
\item protocol_description: A description of the protocol used in the experiment
\item accession_code: The accession code of the experiment
\item alternate_accession_code: An alternate accession code for the experiment
\item source_database: The database from which the experiment was downloaded
\item source_url: The URL from which the experiment was downloaded
\item has_publication: Whether the experiment has a publication associated with it
\item publication_title: The title of the publication associated with the experiment
\item publication_doi: The DOI of the publication associated with the experiment
\item publication_authors: The authors of the publication associated with the experiment
\item pubmed_id: The PubMed ID of the publication associated with the experiment
\item source_first_published: The date the experiment was first published
\item source_last_modified: The date the experiment was last modified
\item submitter_institution: The institution that submitted the experiment
\item last_modified: The date the experiment was last modified
\item created_at: The date the experiment was created
\item organism_names: The names of the organisms in the experiment
\item sample_metadata: The types of metadata available for the samples in the experiment
\item num_total_samples: The total number of samples in the experiment
\item num_processed_samples: The number of samples in the experiment that have been processed
\item num_downloadable_samples: The number of samples in the experiment that are downloadable
}
}
\description{
Experiments are the top level of the refine.bio hierarchy.
They are collections of samples that were processed together.
This function returns metadata giving information about the
experiment, as well as a list of samples that are part of the
experiment.
}
\examples{
expt_details <- rb_experiment("SRP150473")

expt_details$accession_code
expt_details$source_database
head(expt_details$samples)
expt_details$sample_metadata
expt_details$num_total_samples
expt_details$num_processed_samples
expt_details$num_downloadable_samples

}
