% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{rb_experiment_search}
\alias{rb_experiment_search}
\title{Search for experiments}
\usage{
rb_experiment_search(
  id = NULL,
  technology = NULL,
  has_publication = NULL,
  accession_code = NULL,
  alternate_accession_code = NULL,
  platform = NULL,
  organism = NULL,
  downloadable_organism = NULL,
  num_processed_samples = NULL,
  num_downloadable_samples = NULL,
  sample_keywords = NULL,
  ordering = NULL,
  search = NULL,
  pages = Inf,
  limit = 1000,
  .progress = TRUE,
  offset = 0
)
}
\description{
Search for experiments
}
\examples{
experiments <- rb_experiment_search(
  search = "(pediatric) AND (medulloblastoma)",
)

head(experiments)

}
