% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_experiments_read.R
\name{rb_experiments_read}
\alias{rb_experiments_read}
\title{Retrieve details for an experiment given it's accession code}
\usage{
rb_experiments_read(accession_code)
}
\arguments{
\item{accession_code}{string  (required)}
}
\description{
Retrieve details for an experiment given it's accession code
}
\seealso{
Other experiments: 
\code{\link{rb_experiments_list}()}
}
\concept{experiments}
