% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{rb_get_client}
\alias{rb_get_client}
\title{Low level api client accessor}
\usage{
rb_get_client(url = .refinebio_url, renew = FALSE, ...)
}
\arguments{
\item{url}{character(1) the url of the refine bio API root, default
\code{.refinebio_url}}

\item{renew}{logical(1) TRUE to reconnect and rebuild API client,}

\item{...}{passed directly to \code{\link[rapiclient:get_api]{rapiclient::get_api()}}}
}
\description{
This function is an accessor for the
\href{https://api.refine.bio}{refine.bio API}.
On first access, the rapiclient-based client
will be built and returned. On subsequent calls,
the client will be fetched from an environment-based
cache.
}
\examples{

client = RefineBio()

# operations
ops = rapiclient::get_operations(client)



}
\seealso{
\code{\link[rapiclient:get_api]{rapiclient::get_api()}}
}
\concept{low-level-api}
