% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_get_token.R
\name{rb_get_token}
\alias{rb_get_token}
\title{Create or get an active API token for use in downloading from the API}
\usage{
rb_get_token(activate = TRUE, quiet = TRUE)
}
\arguments{
\item{activate}{logical(1) Setting to \code{TRUE} implies agreement with
the refine.bio terms of service for downloading.}

\item{quiet}{logical(1) Setting to \code{FALSE} will message the complete
text of the terms of services.}
}
\description{
The API token mechanism that refine.bio uses is relatively simple.
There is no "authentication" per se, just an acknowledgement of
agreement to terms and conditions of use. The token is only needed
for download of data.
}
\details{
Upon calling this function the first time in a session, if a token
is successfully obtained and activated, it will be stored in the session
in memory. Further calls will use the cached value. A new R session will
result in a new API token.
}
