% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_metadata.R
\name{rb_institution_list}
\alias{rb_institution_list}
\title{All available institutions in refine.bio}
\usage{
rb_institution_list()
}
\description{
All available institutions in refine.bio
}
\examples{
insts <- rb_institution_list()
head(insts)

}
