% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_institutions_list.R
\name{rb_institutions_list}
\alias{rb_institutions_list}
\title{Unpaginated list of all the available "institution" information}
\usage{
rb_institutions_list()
}
\description{
Unpaginated list of all the available "institution" information
}
\concept{institutions}
