% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_jobs_downloader_list.R
\name{rb_jobs_downloader_list}
\alias{rb_jobs_downloader_list}
\title{List of all DownloaderJob}
\usage{
rb_jobs_downloader_list(
  id = NULL,
  downloader_task = NULL,
  ram_amount = NULL,
  num_retries = NULL,
  retried = NULL,
  was_recreated = NULL,
  worker_id = NULL,
  worker_version = NULL,
  batch_job_id = NULL,
  batch_job_queue = NULL,
  failure_reason = NULL,
  success = NULL,
  original_files = NULL,
  start_time = NULL,
  end_time = NULL,
  created_at = NULL,
  last_modified = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL,
  sample_accession_code = NULL
)
}
\arguments{
\item{id}{number}

\item{downloader_task}{string}

\item{ram_amount}{number}

\item{num_retries}{number}

\item{retried}{string}

\item{was_recreated}{string}

\item{worker_id}{string}

\item{worker_version}{string}

\item{batch_job_id}{string}

\item{batch_job_queue}{string}

\item{failure_reason}{string}

\item{success}{string}

\item{original_files}{string}

\item{start_time}{string}

\item{end_time}{string}

\item{created_at}{string}

\item{last_modified}{string}

\item{ordering}{string Which field to use when ordering the results.}

\item{limit}{integer Number of results to return per page.}

\item{offset}{integer The initial index from which to return the results.}

\item{sample_accession_code}{string List the downloader jobs associated with a sample}
}
\description{
List of all DownloaderJob
}
\seealso{
Other jobs: 
\code{\link{rb_jobs_downloader_read}()},
\code{\link{rb_jobs_processor_list}()},
\code{\link{rb_jobs_processor_read}()},
\code{\link{rb_jobs_survey_list}()},
\code{\link{rb_jobs_survey_read}()}
}
\concept{jobs}
