% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_jobs_downloader_read.R
\name{rb_jobs_downloader_read}
\alias{rb_jobs_downloader_read}
\title{Retrieves a DownloaderJob by ID}
\usage{
rb_jobs_downloader_read(id)
}
\arguments{
\item{id}{integer A unique integer value identifying this downloader job. (required)}
}
\description{
Retrieves a DownloaderJob by ID
}
\seealso{
Other jobs: 
\code{\link{rb_jobs_downloader_list}()},
\code{\link{rb_jobs_processor_list}()},
\code{\link{rb_jobs_processor_read}()},
\code{\link{rb_jobs_survey_list}()},
\code{\link{rb_jobs_survey_read}()}
}
\concept{jobs}
