% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_jobs_processor_list.R
\name{rb_jobs_processor_list}
\alias{rb_jobs_processor_list}
\title{List of all ProcessorJobs.}
\usage{
rb_jobs_processor_list(
  id = NULL,
  pipeline_applied = NULL,
  num_retries = NULL,
  retried = NULL,
  worker_id = NULL,
  ram_amount = NULL,
  volume_index = NULL,
  batch_job_queue = NULL,
  worker_version = NULL,
  failure_reason = NULL,
  batch_job_id = NULL,
  success = NULL,
  original_files = NULL,
  datasets = NULL,
  start_time = NULL,
  end_time = NULL,
  created_at = NULL,
  last_modified = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL,
  sample_accession_code = NULL
)
}
\arguments{
\item{id}{number}

\item{pipeline_applied}{string}

\item{num_retries}{number}

\item{retried}{string}

\item{worker_id}{string}

\item{ram_amount}{number}

\item{volume_index}{string}

\item{batch_job_queue}{string}

\item{worker_version}{string}

\item{failure_reason}{string}

\item{batch_job_id}{string}

\item{success}{string}

\item{original_files}{string}

\item{datasets}{string}

\item{start_time}{string}

\item{end_time}{string}

\item{created_at}{string}

\item{last_modified}{string}

\item{ordering}{string Which field to use when ordering the results.}

\item{limit}{integer Number of results to return per page.}

\item{offset}{integer The initial index from which to return the results.}

\item{sample_accession_code}{string List the processor jobs associated with a sample}
}
\description{
List of all ProcessorJobs.
}
\seealso{
Other jobs: 
\code{\link{rb_jobs_downloader_list}()},
\code{\link{rb_jobs_downloader_read}()},
\code{\link{rb_jobs_processor_read}()},
\code{\link{rb_jobs_survey_list}()},
\code{\link{rb_jobs_survey_read}()}
}
\concept{jobs}
