% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_jobs_survey_list.R
\name{rb_jobs_survey_list}
\alias{rb_jobs_survey_list}
\title{List of all SurveyJob.}
\usage{
rb_jobs_survey_list(
  id = NULL,
  source_type = NULL,
  success = NULL,
  batch_job_id = NULL,
  batch_job_queue = NULL,
  ram_amount = NULL,
  start_time = NULL,
  end_time = NULL,
  created_at = NULL,
  last_modified = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{id}{number}

\item{source_type}{string}

\item{success}{string}

\item{batch_job_id}{string}

\item{batch_job_queue}{string}

\item{ram_amount}{number}

\item{start_time}{string}

\item{end_time}{string}

\item{created_at}{string}

\item{last_modified}{string}

\item{ordering}{string Which field to use when ordering the results.}

\item{limit}{integer Number of results to return per page.}

\item{offset}{integer The initial index from which to return the results.}
}
\description{
List of all SurveyJob.
}
\seealso{
Other jobs: 
\code{\link{rb_jobs_downloader_list}()},
\code{\link{rb_jobs_downloader_read}()},
\code{\link{rb_jobs_processor_list}()},
\code{\link{rb_jobs_processor_read}()},
\code{\link{rb_jobs_survey_read}()}
}
\concept{jobs}
