% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_metadata.R
\name{rb_organism_list}
\alias{rb_organism_list}
\title{All available organisms in refine.bio}
\usage{
rb_organism_list()
}
\description{
All available organisms in refine.bio
}
\examples{
orgs <- rb_organism_list()
head(orgs)

}
