% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organisms.R
\name{rb_organisms}
\alias{rb_organisms}
\title{List all available organisms from refine.bio}
\usage{
rb_organisms(client = RefineBio::RefineBio())
}
\value{
a \code{data.frame} with four columns:
\itemize{
\item name
\item taxonomy_id
\item has_compendia
\item has_quantfile_compendia
}
}
\description{
List all available organisms from refine.bio
}
\examples{

rb_orgs = rb_organisms()
head(rb_orgs)

table(rb_orgs$has_compendia)
table(rb_orgs$has_quantfile_compendia)

}
