% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_organisms_read.R
\name{rb_organisms_read}
\alias{rb_organisms_read}
\title{Retrieves an organism by its name.}
\usage{
rb_organisms_read(name)
}
\arguments{
\item{name}{string  (required)}
}
\description{
Retrieves an organism by its name.
}
\seealso{
Other organisms: 
\code{\link{rb_organisms_list}()}
}
\concept{organisms}
