% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_original_files_list.R
\name{rb_original_files_list}
\alias{rb_original_files_list}
\title{List Original Files that are associated with Samples. These are the files we proccess.}
\usage{
rb_original_files_list(
  id = NULL,
  filename = NULL,
  samples = NULL,
  processor_jobs = NULL,
  downloader_jobs = NULL,
  size_in_bytes = NULL,
  sha1 = NULL,
  source_url = NULL,
  is_archive = NULL,
  source_filename = NULL,
  has_raw = NULL,
  created_at = NULL,
  last_modified = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{id}{number}

\item{filename}{string}

\item{samples}{string}

\item{processor_jobs}{string}

\item{downloader_jobs}{string}

\item{size_in_bytes}{number}

\item{sha1}{string}

\item{source_url}{string}

\item{is_archive}{string}

\item{source_filename}{string}

\item{has_raw}{string}

\item{created_at}{string}

\item{last_modified}{string}

\item{ordering}{string Which field to use when ordering the results.}

\item{limit}{integer Number of results to return per page.}

\item{offset}{integer The initial index from which to return the results.}
}
\description{
List Original Files that are associated with Samples. These are the files we proccess.
}
\seealso{
Other original_files: 
\code{\link{rb_original_files_read}()}
}
\concept{original_files}
