% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_original_files_read.R
\name{rb_original_files_read}
\alias{rb_original_files_read}
\title{Retrieves an Original File by its ID}
\usage{
rb_original_files_read(id)
}
\arguments{
\item{id}{integer A unique integer value identifying this original file. (required)}
}
\description{
Retrieves an Original File by its ID
}
\seealso{
Other original_files: 
\code{\link{rb_original_files_list}()}
}
\concept{original_files}
