% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_metadata.R
\name{rb_platform_list}
\alias{rb_platform_list}
\title{All available platforms in refine.bio}
\usage{
rb_platform_list()
}
\value{
\code{data.frame} All available platforms.
}
\description{
All available platforms in refine.bio
}
\examples{
platforms <- rb_platform_list()
head(platforms)

}
\seealso{
Other refine.bio API: 
\code{\link{rb_compendia_list}()},
\code{\link{rb_entity_list}()},
\code{\link{rb_institution_list}()},
\code{\link{rb_organism_list}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{refine.bio API}
