% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{rb_platforms}
\alias{rb_platforms}
\title{List all available platforms from refine.bio}
\usage{
rb_platforms(client = RefineBio::RefineBio())
}
\value{
a 2-column `data.frame`
}
\description{
List all available platforms from refine.bio
}
\examples{

plats = rb_platforms()
head(plats)

}
