% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_platforms_list.R
\name{rb_platforms_list}
\alias{rb_platforms_list}
\title{Unpaginated list of all the available "platform" information}
\usage{
rb_platforms_list()
}
\description{
Unpaginated list of all the available "platform" information
}
\concept{platforms}
