% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refinebio.R
\name{rb_process_and_load}
\alias{rb_process_and_load}
\title{Process and load a dataset}
\usage{
rb_process_and_load(ds, base_path = rb_data_path())
}
\arguments{
\item{ds}{The dataset ID or \code{rb_dataset} object}

\item{base_path}{The path to the directory where the dataset will be
downloaded and extracted.}

\item{cache_result}{If TRUE, cache the result in the base_path.}

\item{overwrite}{If TRUE, overwrite the cached result.}
}
\value{
A list of \code{SummarizedExperiment}s representing the
experiments in the RefineBio downloaded dataset.
}
\description{
This function will process and load a dataset from RefineBio. It is
intended for use with the \code{rb_get_experiment} function.
}
