% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_processors_read.R
\name{rb_processors_read}
\alias{rb_processors_read}
\title{Retrieves a processor by its ID}
\usage{
rb_processors_read(id)
}
\arguments{
\item{id}{integer A unique integer value identifying this processor. (required)}
}
\description{
Retrieves a processor by its ID
}
\seealso{
Other processors: 
\code{\link{rb_processors_list}()}
}
\concept{processors}
