% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_qn_targets_list.R
\name{rb_qn_targets_list}
\alias{rb_qn_targets_list}
\title{This is a list of all of the organisms which have available QN Targets}
\usage{
rb_qn_targets_list()
}
\description{
This is a list of all of the organisms which have available QN Targets
}
\seealso{
Other qn_targets: 
\code{\link{rb_qn_targets_read}()}
}
\concept{qn_targets}
