% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_qn_targets_read.R
\name{rb_qn_targets_read}
\alias{rb_qn_targets_read}
\title{Get a detailed view of the Quantile Normalization file for an organism.}
\usage{
rb_qn_targets_read(organism_name)
}
\arguments{
\item{organism_name}{string Eg \code{DANIO_RERIO}, \code{MUS_MUSCULUS} (required)}
}
\description{
Get a detailed view of the Quantile Normalization file for an organism.
}
\seealso{
Other qn_targets: 
\code{\link{rb_qn_targets_list}()}
}
\concept{qn_targets}
