% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_samples_list.R
\name{rb_samples_list}
\alias{rb_samples_list}
\title{Returns detailed information about Samples}
\usage{
rb_samples_list(
  ordering = NULL,
  title = NULL,
  organism = NULL,
  source_database = NULL,
  source_archive_url = NULL,
  has_raw = NULL,
  platform_name = NULL,
  technology = NULL,
  manufacturer = NULL,
  sex = NULL,
  age = NULL,
  specimen_part = NULL,
  genotype = NULL,
  disease = NULL,
  disease_stage = NULL,
  cell_line = NULL,
  treatment = NULL,
  race = NULL,
  subject = NULL,
  compound = NULL,
  time = NULL,
  is_processed = NULL,
  is_unable_to_be_processed = NULL,
  is_public = NULL,
  limit = NULL,
  offset = NULL,
  dataset_id = NULL,
  experiment_accession_code = NULL,
  accession_codes = NULL
)
}
\arguments{
\item{ordering}{string Which field to use when ordering the results.}

\item{title}{string}

\item{organism}{string}

\item{source_database}{string}

\item{source_archive_url}{string}

\item{has_raw}{string}

\item{platform_name}{string}

\item{technology}{string}

\item{manufacturer}{string}

\item{sex}{string}

\item{age}{number}

\item{specimen_part}{string}

\item{genotype}{string}

\item{disease}{string}

\item{disease_stage}{string}

\item{cell_line}{string}

\item{treatment}{string}

\item{race}{string}

\item{subject}{string}

\item{compound}{string}

\item{time}{string}

\item{is_processed}{string}

\item{is_unable_to_be_processed}{string}

\item{is_public}{string}

\item{limit}{integer Number of results to return per page.}

\item{offset}{integer The initial index from which to return the results.}

\item{dataset_id}{string Filters the result and only returns samples that are added to a dataset.}

\item{experiment_accession_code}{string Filters the result and only returns only the samples associated with an experiment accession code.}

\item{accession_codes}{string Provide a list of sample accession codes separated by commas and the endpoint will only return information about these samples.}
}
\description{
Returns detailed information about Samples
}
\seealso{
Other samples: 
\code{\link{rb_samples_read}()}
}
\concept{samples}
