% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_samples_read.R
\name{rb_samples_read}
\alias{rb_samples_read}
\title{Retrieve the details for a Sample given its accession code}
\usage{
rb_samples_read(accession_code)
}
\arguments{
\item{accession_code}{string  (required)}
}
\description{
Retrieve the details for a Sample given its accession code
}
\seealso{
Other samples: 
\code{\link{rb_samples_list}()}
}
\concept{samples}
