% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{rb_search}
\alias{rb_search}
\title{Perform a search against refine.bio api}
\usage{
rb_search(client, ...)
}
\arguments{
\item{id}{integer(1) the ID to search for}

\item{technology}{character() vector of technologies to search}

\item{has_publiction}{logical(1) filter out records that do not have
associated publications}

\item{accession_code}{character() vector of accessions}

\item{alternate_accession_ccode}{character() vector of alternate accession
codes}

\item{platform}{character() vector of platforms to include}

\item{organism}{character() vector of organisms to include}

\item{downloadable_organism}{character() vector of organisms with data
available for download}

\item{num_processed_samples}{integer() number of processed samples}

\item{sample_keywords}{character() vector of sample keywords to include}

\item{ordering}{character() which field to use when ordering results}

\item{search}{character(1) Search in title, publication_authors,
sample_keywords, publication_title, submitter_institution, description,
accession_code, alternate_accession_code, publication_doi,
pubmed_id, sample_metadata_fields, platform_names.}

\item{limit}{integer(1) number of results to return per page}

\item{offset}{integer(1) the initial offset from with to return the
results}

\item{total}{integer(1) the total number of records to return}
}
\description{
Perform a search against refine.bio api
}
\examples{

client = RefineBio::RefineBio()

sres = rb_search(client, search="medulloblastoma")
sres$count
head(sres$results,4)
lapply(sres$facets,head)


}
