% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_stats_failures_processor_list.R
\name{rb_stats_failures_processor_list}
\alias{rb_stats_failures_processor_list}
\title{API endpoint stats_failures_processor_list}
\usage{
rb_stats_failures_processor_list(range = NULL)
}
\arguments{
\item{range}{string Specify a range from which to calculate the possible options}
}
\description{
API endpoint stats_failures_processor_list
}
\seealso{
Other stats: 
\code{\link{rb_stats_failures_downloader_list}()},
\code{\link{rb_stats_list}()}
}
\concept{stats}
