% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_stats_list.R
\name{rb_stats_list}
\alias{rb_stats_list}
\title{Statistics about the health of the system.}
\usage{
rb_stats_list(range = NULL)
}
\arguments{
\item{range}{string Specify a range from which to calculate the possible options}
}
\description{
Statistics about the health of the system.
}
\seealso{
Other stats: 
\code{\link{rb_stats_failures_downloader_list}()},
\code{\link{rb_stats_failures_processor_list}()}
}
\concept{stats}
