% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_token_create.R
\name{rb_token_create}
\alias{rb_token_create}
\title{This endpoint can be used to create and activate tokens. These tokens can be used
in requests that provide urls to download computed files. Setting \code{is_activated} to
true indicates agreement with refine.bio's \href{https://www.refine.bio/terms}{Terms of Use}
and \href{https://www.refine.bio/privacy}{Privacy Policy}.}
\usage{
rb_token_create(id = NULL, is_activated = NULL, terms_and_conditions = NULL)
}
\arguments{
\item{id}{string}

\item{is_activated}{boolean}

\item{terms_and_conditions}{string}
}
\description{
\if{html}{\out{<div class="sourceCode py">}}\preformatted{import requests
import json
}\if{html}{\out{</div>}}
}
\details{
response = requests.post('https://api.refine.bio/v1/token/')
token_id = response.json()['id']
response = requests.put('https://api.refine.bio/v1/token/' + token_id + '/', json.dumps({'is_activated': True}), headers={'Content-Type': 'application/json'})\preformatted{
The token id needs to be provided in the HTTP request in the API-KEY header.

References
- [https://github.com/AlexsLemonade/refinebio/issues/731]()
- [https://github.com/AlexsLemonade/refinebio-frontend/issues/560]()

['id']: R:'id'
[https://github.com/AlexsLemonade/refinebio/issues/731]: R:https://github.com/AlexsLemonade/refinebio/issues/731
[https://github.com/AlexsLemonade/refinebio-frontend/issues/560]: R:https://github.com/AlexsLemonade/refinebio-frontend/issues/560
}
}
\seealso{
Other token: 
\code{\link{rb_token_read}()},
\code{\link{rb_token_update}()}
}
\concept{token}
