% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_token_read.R
\name{rb_token_read}
\alias{rb_token_read}
\title{Return details about a specific token.}
\usage{
rb_token_read(id)
}
\arguments{
\item{id}{string A UUID string identifying this api token. (required)}
}
\description{
Return details about a specific token.
}
\seealso{
Other token: 
\code{\link{rb_token_create}()},
\code{\link{rb_token_update}()}
}
\concept{token}
