% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_token_update.R
\name{rb_token_update}
\alias{rb_token_update}
\title{This can be used to activate a specific token by sending \code{is_activated: true}.
Setting \code{is_activated} to true indicates agreement with refine.bio's
\href{https://www.refine.bio/terms}{Terms of Use} and
\href{https://www.refine.bio/privacy}{Privacy Policy}.}
\usage{
rb_token_update(id = NULL, is_activated = NULL, terms_and_conditions = NULL)
}
\arguments{
\item{id}{string}

\item{is_activated}{boolean}

\item{terms_and_conditions}{string}
}
\description{
This can be used to activate a specific token by sending \code{is_activated: true}.
Setting \code{is_activated} to true indicates agreement with refine.bio's
\href{https://www.refine.bio/terms}{Terms of Use} and
\href{https://www.refine.bio/privacy}{Privacy Policy}.
}
\seealso{
Other token: 
\code{\link{rb_token_create}()},
\code{\link{rb_token_read}()}
}
\concept{token}
