% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_transcriptome_indices_read.R
\name{rb_transcriptome_indices_read}
\alias{rb_transcriptome_indices_read}
\title{Gets the S3 url associated with the organism and length, along with other metadata about
the transcriptome index we have stored.}
\usage{
rb_transcriptome_indices_read(id)
}
\arguments{
\item{id}{number Transcriptome Index Id eg \code{1} (required)}
}
\description{
Gets the S3 url associated with the organism and length, along with other metadata about
the transcriptome index we have stored.
}
\seealso{
Other transcriptome_indices: 
\code{\link{rb_transcriptome_indices_list}()}
}
\concept{transcriptome_indices}
