% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{rb_wait_for_dataset}
\alias{rb_wait_for_dataset}
\title{Wait for a dataset to become available}
\usage{
rb_wait_for_dataset(dataset, seconds = 10)
}
\arguments{
\item{dataset}{The dataset ID or \code{rb_dataset} object}

\item{seconds}{The number of seconds to wait between checks}
}
\value{
The dataset object when it is available
}
\description{
Wait for a dataset to become available
}
