% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{RefineBio}
\alias{RefineBio}
\title{Low level api client}
\usage{
RefineBio(url = .refinebio_url, ...)
}
\arguments{
\item{url}{character(1) the url of the refine bio API root, default
\code{.refinebio_url}}

\item{...}{passed directly to \code{\link[rapiclient:get_api]{rapiclient::get_api()}}}
}
\description{
This function is an accessor for the
\href{https://api.refine.bio}{refine.bio API}.
}
\examples{

client = RefineBio()

# operations
ops = rapiclient::operations(client)



}
\seealso{
\code{\link[rapiclient:get_api]{rapiclient::get_api()}}
}
\concept{low-level-api}
