% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_r6_client.R
\name{RefineBioClient}
\alias{RefineBioClient}
\title{The entrypoint for the RefineBio API}
\description{
Here we start
List all CompendiaResults with filtering.
}
\concept{compendia
This lists all \code{ComputationalResult}. Each one contains meta-information about the output of a computer process. (Ex Salmon).

This can return valid S3 urls if a valid \href{#tag/token}{token} is sent in the header \code{HTTP_API_KEY}.}
\concept{computational_results
ComputedFiles are representation of files created by refinebio processes.

It's possible to download each one of these files by providing a valid token. To
acquire and activate an API key see the documentation for the \href{#tag/token}{/token} endpoint.
When a valid token is provided the url will be sent back in the field \code{download_url}. Example:

\if{html}{\out{<div class="sourceCode py">}}\preformatted{import requests
import json
headers = \{
    'Content-Type': 'application/json',
    'API-KEY': token_id # requested from /token
\}
requests.get('https://api.refine.bio/v1/computed_files/?id=5796866', \{\}, headers=headers)
}\if{html}{\out{</div>}}

This endpoint can also be used to fetch all the compendia files we have generated with:

\if{html}{\out{<div class="sourceCode">}}\preformatted{GET /computed_files?is_compendia=True&is_public=True
}\if{html}{\out{</div>}}}
\concept{computed_files
View and modify a single Dataset.}
\concept{dataset

Modify an existing Dataset.

In order to begin smashing, an activated API key must be provided in the \code{API-KEY} header field of the request.
To acquire and activate an API key see the documentation for the \href{#tag/token}{/token}
endpoint.

\if{html}{\out{<div class="sourceCode py">}}\preformatted{import requests
import json

params = json.dumps(\{
    'data': data,
    'aggregate_by': 'EXPERIMENT',
    'start': True,
    'email_address': 'refinebio@gmail.com'
\})
headers = \{
    'Content-Type': 'application/json',
    'API-KEY': token_id # requested from /token
\}
requests.put(host + '/v1/dataset/38879729-93c8-436d-9293-b95d3f274741/', params, headers=headers)
}\if{html}{\out{</div>}}}
\concept{dataset
Paginated list of all experiments. Advanced filtering can be done with the \verb{/search} endpoint.}
\concept{experiments
Unpaginated list of all the available "institution" information}
\concept{institutions
List of all DownloaderJob}
\concept{jobs
List of all ProcessorJobs.}
\concept{jobs
List of all SurveyJob.}
\concept{jobs
Paginated list of all the available organisms.}
\concept{organisms
List Original Files that are associated with Samples. These are the files we proccess.}
\concept{original_files
Unpaginated list of all the available "platform" information}
\concept{platforms
List all processors.}
\concept{processors
This is a list of all of the organisms which have available QN Targets}
\concept{qn_targets
Returns detailed information about Samples}
\concept{samples

Use this endpoint to search among the experiments.

This is powered by ElasticSearch, information regarding advanced usages of the
filters can be found in the \href{https://django-elasticsearch-dsl-drf.readthedocs.io/en/0.17.1/filtering_usage_examples.html#filtering}{Django-ES-DSL-DRF docs}

There's an additional field in the response named \code{facets} that contain stats on the number of results per filter type.

Example Requests:

\if{html}{\out{<div class="sourceCode">}}\preformatted{?search=medulloblastoma
?id=1
?search=medulloblastoma&technology=microarray&has_publication=true
?ordering=source_first_published
}\if{html}{\out{</div>}}

This endpoint also accepts POST requests for larger queries. Any of the filters
accepted as query parameters are also accepted in a JSON object in the request
body.

Example Requests (from our tests):

\if{html}{\out{<div class="sourceCode python">}}\preformatted{import requests
import json

headers = \{
    'Content-Type': 'application/json',
\}

# Basic filter
search = \{"accession_code": "GSE123"\}
requests.post(host + '/v1/search/', json.dumps(search), headers=headers)

# __in filter
search = \{"accession_code__in": ["GSE123"]\}
requests.post(host + '/v1/search/', json.dumps(search), headers=headers)

# numeric filter
search = \{"num_downloadable_samples__gt": 0\}
requests.post(host + '/v1/search/', json.dumps(search), headers=headers)
}\if{html}{\out{</div>}}}
\concept{search
Statistics about the health of the system.}
\concept{stats

This endpoint can be used to create and activate tokens. These tokens can be used
in requests that provide urls to download computed files. Setting \code{is_activated} to
true indicates agreement with refine.bio's \href{https://www.refine.bio/terms}{Terms of Use}
and \href{https://www.refine.bio/privacy}{Privacy Policy}.

\if{html}{\out{<div class="sourceCode py">}}\preformatted{import requests
import json

response = requests.post('https://api.refine.bio/v1/token/')
token_id = response.json()['id']
response = requests.put('https://api.refine.bio/v1/token/' + token_id + '/', json.dumps(\{'is_activated': True\}), headers=\{'Content-Type': 'application/json'\})
}\if{html}{\out{</div>}}

The token id needs to be provided in the HTTP request in the API-KEY header.

References
\itemize{
\item \url{https://github.com/AlexsLemonade/refinebio/issues/731}
\item \url{https://github.com/AlexsLemonade/refinebio-frontend/issues/560}
}}
\concept{stats
API endpoint stats_failures_downloader_list}
\concept{stats
API endpoint stats_failures_processor_list}
\concept{token

This can be used to activate a specific token by sending \code{is_activated: true}.
Setting \code{is_activated} to true indicates agreement with refine.bio's
\href{https://www.refine.bio/terms}{Terms of Use} and
\href{https://www.refine.bio/privacy}{Privacy Policy}.}
\concept{token
List all Transcriptome Indices. These are a special type of process result,
necessary for processing other SRA samples.}
\concept{transcriptome_indices}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RefineBioClient-rb_compendia_list}{\code{RefineBioClient$rb_compendia_list()}}
\item \href{#method-RefineBioClient-rb_computational_results_list}{\code{RefineBioClient$rb_computational_results_list()}}
\item \href{#method-RefineBioClient-rb_computed_files_list}{\code{RefineBioClient$rb_computed_files_list()}}
\item \href{#method-RefineBioClient-rb_dataset_create}{\code{RefineBioClient$rb_dataset_create()}}
\item \href{#method-RefineBioClient-rb_dataset_update}{\code{RefineBioClient$rb_dataset_update()}}
\item \href{#method-RefineBioClient-rb_experiments_list}{\code{RefineBioClient$rb_experiments_list()}}
\item \href{#method-RefineBioClient-rb_institutions_list}{\code{RefineBioClient$rb_institutions_list()}}
\item \href{#method-RefineBioClient-rb_jobs_downloader_list}{\code{RefineBioClient$rb_jobs_downloader_list()}}
\item \href{#method-RefineBioClient-rb_jobs_processor_list}{\code{RefineBioClient$rb_jobs_processor_list()}}
\item \href{#method-RefineBioClient-rb_jobs_survey_list}{\code{RefineBioClient$rb_jobs_survey_list()}}
\item \href{#method-RefineBioClient-rb_organisms_list}{\code{RefineBioClient$rb_organisms_list()}}
\item \href{#method-RefineBioClient-rb_original_files_list}{\code{RefineBioClient$rb_original_files_list()}}
\item \href{#method-RefineBioClient-rb_platforms_list}{\code{RefineBioClient$rb_platforms_list()}}
\item \href{#method-RefineBioClient-rb_processors_list}{\code{RefineBioClient$rb_processors_list()}}
\item \href{#method-RefineBioClient-rb_qn_targets_list}{\code{RefineBioClient$rb_qn_targets_list()}}
\item \href{#method-RefineBioClient-rb_samples_list}{\code{RefineBioClient$rb_samples_list()}}
\item \href{#method-RefineBioClient-rb_search_list}{\code{RefineBioClient$rb_search_list()}}
\item \href{#method-RefineBioClient-rb_stats_list}{\code{RefineBioClient$rb_stats_list()}}
\item \href{#method-RefineBioClient-rb_stats_failures_downloader_list}{\code{RefineBioClient$rb_stats_failures_downloader_list()}}
\item \href{#method-RefineBioClient-rb_stats_failures_processor_list}{\code{RefineBioClient$rb_stats_failures_processor_list()}}
\item \href{#method-RefineBioClient-rb_token_create}{\code{RefineBioClient$rb_token_create()}}
\item \href{#method-RefineBioClient-rb_token_update}{\code{RefineBioClient$rb_token_update()}}
\item \href{#method-RefineBioClient-rb_transcriptome_indices_list}{\code{RefineBioClient$rb_transcriptome_indices_list()}}
\item \href{#method-RefineBioClient-clone}{\code{RefineBioClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_compendia_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_compendia_list}{}}}
\subsection{Method \code{rb_compendia_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_compendia_list(
  primary_organism__name = NULL,
  compendium_version = NULL,
  quant_sf_only = NULL,
  result__id = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL,
  latest_version = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{primary_organism__name}}{string}

\item{\code{compendium_version}}{number}

\item{\code{quant_sf_only}}{boolean \code{True} for RNA-seq Sample Compendium results or \code{False} for quantile normalized.}

\item{\code{result__id}}{number}

\item{\code{ordering}}{string Which field to use when ordering the results.}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}

\item{\code{latest_version}}{boolean \code{True} will only return the highest \code{compendium_version} for each primary_organism.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_computational_results_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_computational_results_list}{}}}
\subsection{Method \code{rb_computational_results_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_computational_results_list(
  processor__id = NULL,
  limit = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{processor__id}}{number}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_computed_files_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_computed_files_list}{}}}
\subsection{Method \code{rb_computed_files_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_computed_files_list(
  id = NULL,
  samples = NULL,
  is_qn_target = NULL,
  is_smashable = NULL,
  is_qc = NULL,
  is_compendia = NULL,
  quant_sf_only = NULL,
  svd_algorithm = NULL,
  compendium_version = NULL,
  created_at = NULL,
  last_modified = NULL,
  result__id = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{number}

\item{\code{samples}}{string}

\item{\code{is_qn_target}}{string}

\item{\code{is_smashable}}{string}

\item{\code{is_qc}}{string}

\item{\code{is_compendia}}{string}

\item{\code{quant_sf_only}}{string}

\item{\code{svd_algorithm}}{string}

\item{\code{compendium_version}}{number}

\item{\code{created_at}}{string}

\item{\code{last_modified}}{string}

\item{\code{result__id}}{number}

\item{\code{ordering}}{string Which field to use when ordering the results.}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_dataset_create"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_dataset_create}{}}}
\subsection{Method \code{rb_dataset_create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_dataset_create(
  id = NULL,
  data = NULL,
  aggregate_by = NULL,
  scale_by = NULL,
  is_processing = NULL,
  is_processed = NULL,
  is_available = NULL,
  has_email = NULL,
  email_address = NULL,
  email_ccdl_ok = NULL,
  notify_me = NULL,
  expires_on = NULL,
  s3_bucket = NULL,
  s3_key = NULL,
  success = NULL,
  failure_reason = NULL,
  created_at = NULL,
  last_modified = NULL,
  start = NULL,
  size_in_bytes = NULL,
  sha1 = NULL,
  quantile_normalize = NULL,
  quant_sf_only = NULL,
  svd_algorithm = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{string}

\item{\code{data}}{object This is a dictionary where the keys are experiment accession codes and the values are lists with sample accession codes. Eg: \verb{\{'E-ABC-1': ['SAMP1', 'SAMP2']\}}}

\item{\code{aggregate_by}}{string Specifies how samples are \href{http://docs.refine.bio/en/latest/main_text.html#aggregations}{aggregated}.}

\item{\code{scale_by}}{string Specifies options for \href{http://docs.refine.bio/en/latest/main_text.html#transformations}{transformations}.}

\item{\code{is_processing}}{boolean}

\item{\code{is_processed}}{boolean}

\item{\code{is_available}}{boolean}

\item{\code{has_email}}{string}

\item{\code{email_address}}{string}

\item{\code{email_ccdl_ok}}{boolean}

\item{\code{notify_me}}{boolean}

\item{\code{expires_on}}{string}

\item{\code{s3_bucket}}{string}

\item{\code{s3_key}}{string}

\item{\code{success}}{boolean}

\item{\code{failure_reason}}{string}

\item{\code{created_at}}{string}

\item{\code{last_modified}}{string}

\item{\code{start}}{boolean}

\item{\code{size_in_bytes}}{integer Contains the size in bytes of the processed dataset.}

\item{\code{sha1}}{string}

\item{\code{quantile_normalize}}{boolean Part of the advanced options. Allows \href{http://docs.refine.bio/en/latest/faq.html#what-does-it-mean-to-skip-quantile-normalization-for-rna-seq-samples}{skipping quantile normalization} for RNA-Seq samples.}

\item{\code{quant_sf_only}}{boolean Include only quant.sf files in the generated dataset.}

\item{\code{svd_algorithm}}{string Specifies choice of SVD algorithm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_dataset_update"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_dataset_update}{}}}
\subsection{Method \code{rb_dataset_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_dataset_update(
  id = NULL,
  data = NULL,
  aggregate_by = NULL,
  scale_by = NULL,
  is_processing = NULL,
  is_processed = NULL,
  is_available = NULL,
  has_email = NULL,
  email_address = NULL,
  email_ccdl_ok = NULL,
  notify_me = NULL,
  expires_on = NULL,
  s3_bucket = NULL,
  s3_key = NULL,
  success = NULL,
  failure_reason = NULL,
  created_at = NULL,
  last_modified = NULL,
  start = NULL,
  size_in_bytes = NULL,
  sha1 = NULL,
  quantile_normalize = NULL,
  quant_sf_only = NULL,
  svd_algorithm = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{string}

\item{\code{data}}{object This is a dictionary where the keys are experiment accession codes and the values are lists with sample accession codes. Eg: \verb{\{'E-ABC-1': ['SAMP1', 'SAMP2']\}}}

\item{\code{aggregate_by}}{string Specifies how samples are \href{http://docs.refine.bio/en/latest/main_text.html#aggregations}{aggregated}.}

\item{\code{scale_by}}{string Specifies options for \href{http://docs.refine.bio/en/latest/main_text.html#transformations}{transformations}.}

\item{\code{is_processing}}{boolean}

\item{\code{is_processed}}{boolean}

\item{\code{is_available}}{boolean}

\item{\code{has_email}}{string}

\item{\code{email_address}}{string}

\item{\code{email_ccdl_ok}}{boolean}

\item{\code{notify_me}}{boolean}

\item{\code{expires_on}}{string}

\item{\code{s3_bucket}}{string}

\item{\code{s3_key}}{string}

\item{\code{success}}{boolean}

\item{\code{failure_reason}}{string}

\item{\code{created_at}}{string}

\item{\code{last_modified}}{string}

\item{\code{start}}{boolean}

\item{\code{size_in_bytes}}{integer Contains the size in bytes of the processed dataset.}

\item{\code{sha1}}{string}

\item{\code{quantile_normalize}}{boolean Part of the advanced options. Allows \href{http://docs.refine.bio/en/latest/faq.html#what-does-it-mean-to-skip-quantile-normalization-for-rna-seq-samples}{skipping quantile normalization} for RNA-Seq samples.}

\item{\code{quant_sf_only}}{boolean Include only quant.sf files in the generated dataset.}

\item{\code{svd_algorithm}}{string Specifies choice of SVD algorithm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_experiments_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_experiments_list}{}}}
\subsection{Method \code{rb_experiments_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_experiments_list(
  title = NULL,
  description = NULL,
  accession_code = NULL,
  alternate_accession_code = NULL,
  source_database = NULL,
  source_url = NULL,
  has_publication = NULL,
  publication_title = NULL,
  publication_doi = NULL,
  pubmed_id = NULL,
  organisms = NULL,
  submitter_institution = NULL,
  created_at = NULL,
  last_modified = NULL,
  source_first_published = NULL,
  source_last_modified = NULL,
  limit = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{string}

\item{\code{description}}{string}

\item{\code{accession_code}}{string}

\item{\code{alternate_accession_code}}{string}

\item{\code{source_database}}{string}

\item{\code{source_url}}{string}

\item{\code{has_publication}}{string}

\item{\code{publication_title}}{string}

\item{\code{publication_doi}}{string}

\item{\code{pubmed_id}}{string}

\item{\code{organisms}}{string}

\item{\code{submitter_institution}}{string}

\item{\code{created_at}}{string}

\item{\code{last_modified}}{string}

\item{\code{source_first_published}}{string}

\item{\code{source_last_modified}}{string}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_institutions_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_institutions_list}{}}}
\subsection{Method \code{rb_institutions_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_institutions_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_jobs_downloader_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_jobs_downloader_list}{}}}
\subsection{Method \code{rb_jobs_downloader_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_jobs_downloader_list(
  id = NULL,
  downloader_task = NULL,
  ram_amount = NULL,
  num_retries = NULL,
  retried = NULL,
  was_recreated = NULL,
  worker_id = NULL,
  worker_version = NULL,
  batch_job_id = NULL,
  batch_job_queue = NULL,
  failure_reason = NULL,
  success = NULL,
  original_files = NULL,
  start_time = NULL,
  end_time = NULL,
  created_at = NULL,
  last_modified = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL,
  sample_accession_code = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{number}

\item{\code{downloader_task}}{string}

\item{\code{ram_amount}}{number}

\item{\code{num_retries}}{number}

\item{\code{retried}}{string}

\item{\code{was_recreated}}{string}

\item{\code{worker_id}}{string}

\item{\code{worker_version}}{string}

\item{\code{batch_job_id}}{string}

\item{\code{batch_job_queue}}{string}

\item{\code{failure_reason}}{string}

\item{\code{success}}{string}

\item{\code{original_files}}{string}

\item{\code{start_time}}{string}

\item{\code{end_time}}{string}

\item{\code{created_at}}{string}

\item{\code{last_modified}}{string}

\item{\code{ordering}}{string Which field to use when ordering the results.}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}

\item{\code{sample_accession_code}}{string List the downloader jobs associated with a sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_jobs_processor_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_jobs_processor_list}{}}}
\subsection{Method \code{rb_jobs_processor_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_jobs_processor_list(
  id = NULL,
  pipeline_applied = NULL,
  num_retries = NULL,
  retried = NULL,
  worker_id = NULL,
  ram_amount = NULL,
  volume_index = NULL,
  batch_job_queue = NULL,
  worker_version = NULL,
  failure_reason = NULL,
  batch_job_id = NULL,
  success = NULL,
  original_files = NULL,
  datasets = NULL,
  start_time = NULL,
  end_time = NULL,
  created_at = NULL,
  last_modified = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL,
  sample_accession_code = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{number}

\item{\code{pipeline_applied}}{string}

\item{\code{num_retries}}{number}

\item{\code{retried}}{string}

\item{\code{worker_id}}{string}

\item{\code{ram_amount}}{number}

\item{\code{volume_index}}{string}

\item{\code{batch_job_queue}}{string}

\item{\code{worker_version}}{string}

\item{\code{failure_reason}}{string}

\item{\code{batch_job_id}}{string}

\item{\code{success}}{string}

\item{\code{original_files}}{string}

\item{\code{datasets}}{string}

\item{\code{start_time}}{string}

\item{\code{end_time}}{string}

\item{\code{created_at}}{string}

\item{\code{last_modified}}{string}

\item{\code{ordering}}{string Which field to use when ordering the results.}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}

\item{\code{sample_accession_code}}{string List the processor jobs associated with a sample}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_jobs_survey_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_jobs_survey_list}{}}}
\subsection{Method \code{rb_jobs_survey_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_jobs_survey_list(
  id = NULL,
  source_type = NULL,
  success = NULL,
  batch_job_id = NULL,
  batch_job_queue = NULL,
  ram_amount = NULL,
  start_time = NULL,
  end_time = NULL,
  created_at = NULL,
  last_modified = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{number}

\item{\code{source_type}}{string}

\item{\code{success}}{string}

\item{\code{batch_job_id}}{string}

\item{\code{batch_job_queue}}{string}

\item{\code{ram_amount}}{number}

\item{\code{start_time}}{string}

\item{\code{end_time}}{string}

\item{\code{created_at}}{string}

\item{\code{last_modified}}{string}

\item{\code{ordering}}{string Which field to use when ordering the results.}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_organisms_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_organisms_list}{}}}
\subsection{Method \code{rb_organisms_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_organisms_list(
  has_compendia = NULL,
  has_quantfile_compendia = NULL,
  limit = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{has_compendia}}{string}

\item{\code{has_quantfile_compendia}}{string}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_original_files_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_original_files_list}{}}}
\subsection{Method \code{rb_original_files_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_original_files_list(
  id = NULL,
  filename = NULL,
  samples = NULL,
  processor_jobs = NULL,
  downloader_jobs = NULL,
  size_in_bytes = NULL,
  sha1 = NULL,
  source_url = NULL,
  is_archive = NULL,
  source_filename = NULL,
  has_raw = NULL,
  created_at = NULL,
  last_modified = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{number}

\item{\code{filename}}{string}

\item{\code{samples}}{string}

\item{\code{processor_jobs}}{string}

\item{\code{downloader_jobs}}{string}

\item{\code{size_in_bytes}}{number}

\item{\code{sha1}}{string}

\item{\code{source_url}}{string}

\item{\code{is_archive}}{string}

\item{\code{source_filename}}{string}

\item{\code{has_raw}}{string}

\item{\code{created_at}}{string}

\item{\code{last_modified}}{string}

\item{\code{ordering}}{string Which field to use when ordering the results.}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_platforms_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_platforms_list}{}}}
\subsection{Method \code{rb_platforms_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_platforms_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_processors_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_processors_list}{}}}
\subsection{Method \code{rb_processors_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_processors_list(limit = NULL, offset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_qn_targets_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_qn_targets_list}{}}}
\subsection{Method \code{rb_qn_targets_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_qn_targets_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_samples_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_samples_list}{}}}
\subsection{Method \code{rb_samples_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_samples_list(
  ordering = NULL,
  title = NULL,
  organism = NULL,
  source_database = NULL,
  source_archive_url = NULL,
  has_raw = NULL,
  platform_name = NULL,
  technology = NULL,
  manufacturer = NULL,
  sex = NULL,
  age = NULL,
  specimen_part = NULL,
  genotype = NULL,
  disease = NULL,
  disease_stage = NULL,
  cell_line = NULL,
  treatment = NULL,
  race = NULL,
  subject = NULL,
  compound = NULL,
  time = NULL,
  is_processed = NULL,
  is_unable_to_be_processed = NULL,
  is_public = NULL,
  limit = NULL,
  offset = NULL,
  dataset_id = NULL,
  experiment_accession_code = NULL,
  accession_codes = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ordering}}{string Which field to use when ordering the results.}

\item{\code{title}}{string}

\item{\code{organism}}{string}

\item{\code{source_database}}{string}

\item{\code{source_archive_url}}{string}

\item{\code{has_raw}}{string}

\item{\code{platform_name}}{string}

\item{\code{technology}}{string}

\item{\code{manufacturer}}{string}

\item{\code{sex}}{string}

\item{\code{age}}{number}

\item{\code{specimen_part}}{string}

\item{\code{genotype}}{string}

\item{\code{disease}}{string}

\item{\code{disease_stage}}{string}

\item{\code{cell_line}}{string}

\item{\code{treatment}}{string}

\item{\code{race}}{string}

\item{\code{subject}}{string}

\item{\code{compound}}{string}

\item{\code{time}}{string}

\item{\code{is_processed}}{string}

\item{\code{is_unable_to_be_processed}}{string}

\item{\code{is_public}}{string}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}

\item{\code{dataset_id}}{string Filters the result and only returns samples that are added to a dataset.}

\item{\code{experiment_accession_code}}{string Filters the result and only returns only the samples associated with an experiment accession code.}

\item{\code{accession_codes}}{string Provide a list of sample accession codes separated by commas and the endpoint will only return information about these samples.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_search_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_search_list}{}}}
\subsection{Method \code{rb_search_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_search_list(
  id = NULL,
  technology = NULL,
  has_publication = NULL,
  accession_code = NULL,
  alternate_accession_code = NULL,
  platform = NULL,
  organism = NULL,
  downloadable_organism = NULL,
  num_processed_samples = NULL,
  num_downloadable_samples = NULL,
  sample_keywords = NULL,
  ordering = NULL,
  search = NULL,
  limit = NULL,
  offset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{number}

\item{\code{technology}}{string Allows filtering the results by technology, can have multiple values. Eg: \code{?technology=microarray&technology=rna-seq}}

\item{\code{has_publication}}{string Filter the results that have associated publications with \code{?has_publication=true}}

\item{\code{accession_code}}{string Allows filtering the results by accession code, can have multiple values. Eg: \code{?accession_code=microarray&accession_code=rna-seq}}

\item{\code{alternate_accession_code}}{string}

\item{\code{platform}}{string Allows filtering the results by platform, this parameter can have multiple values.}

\item{\code{organism}}{string Allows filtering the results by organism, this parameter can have multiple values.}

\item{\code{downloadable_organism}}{string}

\item{\code{num_processed_samples}}{number Use ElasticSearch queries to specify the number of processed samples of the results}

\item{\code{num_downloadable_samples}}{number}

\item{\code{sample_keywords}}{string}

\item{\code{ordering}}{string Which field from to use when ordering the results.}

\item{\code{search}}{string Search in title, publication_authors, sample_keywords, publication_title, submitter_institution, description, accession_code, alternate_accession_code, publication_doi, pubmed_id, sample_metadata_fields, platform_names.}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_stats_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_stats_list}{}}}
\subsection{Method \code{rb_stats_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_stats_list(range = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{range}}{string Specify a range from which to calculate the possible options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_stats_failures_downloader_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_stats_failures_downloader_list}{}}}
\subsection{Method \code{rb_stats_failures_downloader_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_stats_failures_downloader_list(range = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{range}}{string Specify a range from which to calculate the possible options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_stats_failures_processor_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_stats_failures_processor_list}{}}}
\subsection{Method \code{rb_stats_failures_processor_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_stats_failures_processor_list(range = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{range}}{string Specify a range from which to calculate the possible options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_token_create"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_token_create}{}}}
\subsection{Method \code{rb_token_create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_token_create(
  id = NULL,
  is_activated = NULL,
  terms_and_conditions = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{string}

\item{\code{is_activated}}{boolean}

\item{\code{terms_and_conditions}}{string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_token_update"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_token_update}{}}}
\subsection{Method \code{rb_token_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_token_update(
  id = NULL,
  is_activated = NULL,
  terms_and_conditions = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{string}

\item{\code{is_activated}}{boolean}

\item{\code{terms_and_conditions}}{string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-rb_transcriptome_indices_list"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-rb_transcriptome_indices_list}{}}}
\subsection{Method \code{rb_transcriptome_indices_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$rb_transcriptome_indices_list(
  salmon_version = NULL,
  index_type = NULL,
  ordering = NULL,
  limit = NULL,
  offset = NULL,
  organism__name = NULL,
  length = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{salmon_version}}{string Eg. \verb{salmon 0.13.1}}

\item{\code{index_type}}{string Eg. \code{TRANSCRIPTOME_LONG}}

\item{\code{ordering}}{string Which field to use when ordering the results.}

\item{\code{limit}}{integer Number of results to return per page.}

\item{\code{offset}}{integer The initial index from which to return the results.}

\item{\code{organism__name}}{string Organism name. Eg. \code{MUS_MUSCULUS}}

\item{\code{length}}{string Short hand for \code{index_type} Eg. \code{short} or \code{long}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
