% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_results.R
\name{RefineBioResults}
\alias{RefineBioResults}
\title{This is my RefineBioResults class}
\description{
class description
}
\examples{
r <- RefineBioResults$new(endpoint = "/experiments", method = "get")
r$next_page()$results
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the person}

\item{\code{hair}}{Hair colour}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the person}

\item{\code{hair}}{Hair colour}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RefineBioResults-new}{\code{RefineBioResults$new()}}
\item \href{#method-RefineBioResults-reset_cursor}{\code{RefineBioResults$reset_cursor()}}
\item \href{#method-RefineBioResults-has_next}{\code{RefineBioResults$has_next()}}
\item \href{#method-RefineBioResults-next_results}{\code{RefineBioResults$next_results()}}
\item \href{#method-RefineBioResults-all_results}{\code{RefineBioResults$all_results()}}
\item \href{#method-RefineBioResults-clone}{\code{RefineBioResults$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioResults-new"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioResults-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioResults$new(
  query = NULL,
  body = NULL,
  method = NULL,
  endpoint = NULL,
  limit = 50,
  offset = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioResults-reset_cursor"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioResults-reset_cursor}{}}}
\subsection{Method \code{reset_cursor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioResults$reset_cursor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioResults-has_next"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioResults-has_next}{}}}
\subsection{Method \code{has_next()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioResults$has_next()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioResults-next_results"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioResults-next_results}{}}}
\subsection{Method \code{next_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioResults$next_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioResults-all_results"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioResults-all_results}{}}}
\subsection{Method \code{all_results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioResults$all_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RefineBioResults-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RefineBioResults-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RefineBioResults$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
