% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_state.R
\name{set_rb_email_address}
\alias{set_rb_email_address}
\title{set the email address for the current session}
\usage{
set_rb_email_address(email_address)
}
\arguments{
\item{email_address}{character string}
}
\description{
set the email address for the current session
}
\examples{
\dontrun{
set_rb_email_address("happy_rb_user@example.com")
}

}
