% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{.facet_to_data_frame}
\alias{.facet_to_data_frame}
\title{Turn facet results into data.frame}
\usage{
.facet_to_data_frame(facet_list)
}
\arguments{
\item{facet_list}{list() with names representing facets and values
for each element representing the count of such facet in the
search result}
}
\value{
\code{data.frame} with columns
\itemize{
\item category
\item count
}
}
\description{
Turn facet results into data.frame
}
\keyword{internal}
