% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refinebio.R
\name{rb_get_experiment}
\alias{rb_get_experiment}
\title{Get one or more experiments (GSE, SRP, etc) from RefineBio}
\usage{
rb_get_experiment(experiment)
}
\arguments{
\item{experiment}{The experiment ID or \code{rb_experiment} object}
}
\value{
A list of \code{SummarizedExperiment}s representing the
experiments in the RefineBio downloaded dataset.
}
\description{
This is the main function for interacting with RefineBio. It will
download, extract, and load the data from RefineBio for the given
experiment(s).
}
\examples{
\dontrun{
# Get a single experiment
expt <- rb_get_experiment("GSE1133")

# Get multiple experiments
expts <- rb_get_experiment(c("GSE1133", "GSE11331"))
}
}
\author{
Sean Davis <seandavi.gmail.com>
}
