% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{rb_get_token}
\alias{rb_get_token}
\title{Create or get an active API token for use in downloading from the API}
\usage{
rb_get_token()
}
\description{
The API token mechanism that refine.bio uses is relatively simple.
There is no "authentication" per se, just an acknowledgement of
agreement to terms and conditions of use. The token is only needed
for download of data. Other functionality will work fine without it.
}
\details{
Upon calling this function the first time in a session, if a token
is successfully obtained and activated, it will be stored in the session
in memory. Further calls will use the cached value. A new R session will
result in a new API token.
}
