% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_metadata.R
\name{rb_institution_list}
\alias{rb_institution_list}
\title{All available institutions in refine.bio}
\usage{
rb_institution_list()
}
\value{
\code{data.frame} All available submitter institutions.
}
\description{
All available institutions in refine.bio
}
\examples{
insts <- rb_institution_list()
head(insts)

}
\seealso{
Other refine.bio API: 
\code{\link{rb_compendia_list}()},
\code{\link{rb_entity_list}()},
\code{\link{rb_organism_list}()},
\code{\link{rb_platform_list}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{refine.bio API}
