% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_metadata.R
\name{rb_organism_list}
\alias{rb_organism_list}
\title{All available organisms in refine.bio}
\usage{
rb_organism_list()
}
\value{
\code{data.frame} All available organisms.
}
\description{
All available organisms in refine.bio
}
\examples{
orgs <- rb_organism_list()
head(orgs)

}
\seealso{
Other refine.bio API: 
\code{\link{rb_compendia_list}()},
\code{\link{rb_entity_list}()},
\code{\link{rb_institution_list}()},
\code{\link{rb_platform_list}()}
}
\author{
Sean Davis \href{mailto:seandavi@gmail.com}{seandavi@gmail.com}
}
\concept{refine.bio API}
