% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_metadata.R
\name{rb_platform_list}
\alias{rb_platform_list}
\title{All available platforms in refine.bio}
\usage{
rb_platform_list()
}
\description{
All available platforms in refine.bio
}
\examples{
platforms <- rb_platform_list()
head(platforms)

}
