% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{RefineBio}
\alias{RefineBio}
\title{Low level api client}
\usage{
RefineBio(url = .refinebio_url, ...)
}
\arguments{
\item{url}{character(1) the url of the refine bio API root, default
`.refinebio_url`}

\item{...}{passed directly to [rapiclient::get_api()]}
}
\description{
This function is an accessor for the
[refine.bio API](https://api.refine.bio).
}
\examples{

client = RefineBio()

# operations
ops = rapiclient::operations(client)



}
\seealso{
[rapiclient::get_api()]
}
\concept{low-level-api}
